// types/zod/outputTypeSchemas/taxi_ordersFindManyArgsSchema.ts
export const taxi_ordersSelectSchema: z.ZodType<Prisma.taxi_ordersSelect> = z
	.object({
		order_id: z.boolean().optional(),
		user_id: z.boolean().optional(),
		business_users_id: z.boolean().optional(),
		business_clients_id: z.boolean().optional(),
		driver_id: z.boolean().optional(),
		vehicle_id: z.boolean().optional(),
		route: z.boolean().optional(),
		pickup_location: z.boolean().optional(),
		delivery_location: z.boolean().optional(),
		payment: z.boolean().optional(),
		estimates: z.boolean().optional(),
		timeline: z.boolean().optional(),
		preferences: z.boolean().optional(),
		status: z.boolean().optional(),
		last_sent_at: z.boolean().optional(),
		created_at: z.boolean().optional(),
		updated_at: z.boolean().optional(),
		business_id: z.boolean().optional(),
		telephone: z.boolean().optional(),
		first_name: z.boolean().optional(),
		last_name: z.boolean().optional(),
		cancellation_reason: z.boolean().optional(),
		find_drivers_attempts: z.boolean().optional(),
		is_scheduled: z.boolean().optional(),
		parent_order_id: z.boolean().optional(),
		type: z.boolean().optional(),
		subtype: z.boolean().optional(),
		flags: z.boolean().optional(),
		cargo_preferences: z.boolean().optional(),
		customer_note: z.boolean().optional(),
		parent_user_type: z.boolean().optional(),
		creating_user_id: z.boolean().optional(),
		allow_credits_usage: z.boolean().optional(),
		order_number: z.boolean().optional(),
		business: z.union([z.boolean(), z.lazy(() => businessArgsSchema)]).optional(),
		driver: z.union([z.boolean(), z.lazy(() => driversArgsSchema)]).optional(),
		vehicle: z.union([z.boolean(), z.lazy(() => vehiclesArgsSchema)]).optional(),
		user: z.union([z.boolean(), z.lazy(() => usersArgsSchema)]).optional(),
		business_users: z.union([z.boolean(), z.lazy(() => business_usersArgsSchema)]).optional(),
		business_clients: z.union([z.boolean(), z.lazy(() => business_clientsArgsSchema)]).optional(),
		history: z.union([z.boolean(), z.lazy(() => taxi_order_sentFindManyArgsSchema)]).optional(),
		parent_order: z.union([z.boolean(), z.lazy(() => taxi_ordersArgsSchema)]).optional(),
		grouped_orders: z.union([z.boolean(), z.lazy(() => taxi_ordersFindManyArgsSchema)]).optional(),
		driver_history_locations: z
			.union([z.boolean(), z.lazy(() => driver_history_locationsFindManyArgsSchema)])
			.optional(),
		wallet_transfer: z.union([z.boolean(), z.lazy(() => wallet_transfer_historyFindManyArgsSchema)]).optional(),
		transactions: z.union([z.boolean(), z.lazy(() => transactionsFindManyArgsSchema)]).optional(),
		cashback: z.union([z.boolean(), z.lazy(() => cashbackFindManyArgsSchema)]).optional(),
		scoring_points: z.union([z.boolean(), z.lazy(() => scoring_pointsFindManyArgsSchema)]).optional(),
		late_events: z.union([z.boolean(), z.lazy(() => late_eventsFindManyArgsSchema)]).optional(),
		_count: z.union([z.boolean(), z.lazy(() => Taxi_ordersCountOutputTypeArgsSchema)]).optional(),
	})
	.strict();

export const taxi_ordersFindManyArgsSchema: z.ZodType<Prisma.taxi_ordersFindManyArgs> = z
	.object({
		select: taxi_ordersSelectSchema.optional(),
		include: taxi_ordersIncludeSchema.optional(),
		where: taxi_ordersWhereInputSchema.optional(),
		orderBy: z
			.union([taxi_ordersOrderByWithRelationInputSchema.array(), taxi_ordersOrderByWithRelationInputSchema])
			.optional(),
		cursor: taxi_ordersWhereUniqueInputSchema.optional(),
		take: z.number().optional(),
		skip: z.number().optional(),
		distinct: z.union([Taxi_ordersScalarFieldEnumSchema, Taxi_ordersScalarFieldEnumSchema.array()]).optional(),
	})
	.strict();